adjust lines
