+++ apps/strength/charts/lib/chartUtils.ts
@@ -1,4 +1,4 @@
-import { LineData, Time } from 'lightweight-charts'
+import { LineData, Time, ISeriesApi } from 'lightweight-charts'
 import { StrengthRowGet } from '@lib/common/sql/strength'
 
 /**
@@ -92,7 +92,6 @@ export const calculateTimeRange = (
   return null
 }
 
-
 /**
  * Get nearest series value at a specific time using binary search
  * If multiple intervals are provided, average their values
@@ -167,4 +166,154 @@ export const getNearestSeriesValueAtTime = (
 
   // Return null if no valid values found, otherwise return average
   return count > 0 ? sum / count : null
+}
+
+/**
+ * Efficiently update chart series data using update() for appends or setData() for full refreshes.
+ *
+ * IMPORTANT: lightweight-charts update() can ONLY:
+ * 1. Update the LAST point in the series (if same timestamp)
+ * 2. Append a new point that comes AFTER the last point
+ *
+ * It CANNOT update points in the middle of the series.
+ * For any other changes, we must use setData().
+ *
+ * @param series - The chart series to update
+ * @param currentData - New data array
+ * @param prevData - Previous data array (for comparison)
+ * @returns true if data was updated, false if no changes detected
+ */
+export const updateSeriesEfficiently = (
+  series: ISeriesApi<'Line'>,
+  currentData: LineData[],
+  prevData: LineData[] | null
+): boolean => {
+  if (currentData.length === 0) return false
+
+  // First load or no previous data: use setData
+  if (!prevData || prevData.length === 0) {
+    series.setData(currentData)
+    return true
+  }
+
+  const lengthDiff = currentData.length - prevData.length
+
+  // Check if this is a simple append (new points at the end only)
+  if (lengthDiff > 0 && lengthDiff <= 10) {
+    // Verify that existing data hasn't changed (timestamps match)
+    // We only check the last few points of the existing data for performance
+    const checkCount = Math.min(5, prevData.length)
+    let existingDataUnchanged = true
+
+    for (let i = 0; i < checkCount; i++) {
+      const idx = prevData.length - 1 - i
+      const curr = currentData[idx]
+      const prev = prevData[idx]
+
+      // If timestamps differ or values significantly differ, data changed
+      if (
+        !curr ||
+        !prev ||
+        curr.time !== prev.time ||
+        Math.abs(curr.value - prev.value) > 0.0001
+      ) {
+        existingDataUnchanged = false
+        break
+      }
+    }
+
+    // Only use update() if existing data is unchanged
+    // (meaning we're just appending new points)
+    if (existingDataUnchanged) {
+      try {
+        // Append only the new points
+        for (let i = prevData.length; i < currentData.length; i++) {
+          const point = currentData[i]
+          if (point) {
+            series.update(point)
+          }
+        }
+        return true
+      } catch (e) {
+        // If update fails for any reason, fall back to setData
+        console.warn('update() failed, falling back to setData:', e)
+        series.setData(currentData)
+        return true
+      }
+    }
+  }
+
+  // Same length - check if only the LAST point changed
+  if (lengthDiff === 0) {
+    const lastCurr = currentData[currentData.length - 1]
+    const lastPrev = prevData[prevData.length - 1]
+
+    // If only the last point changed (same timestamp, different value)
+    if (
+      lastCurr &&
+      lastPrev &&
+      lastCurr.time === lastPrev.time &&
+      Math.abs(lastCurr.value - lastPrev.value) > 0.0001
+    ) {
+      // Check if everything else is the same
+      let onlyLastChanged = true
+      const checkCount = Math.min(5, currentData.length - 1)
+
+      for (let i = 0; i < checkCount; i++) {
+        const idx = currentData.length - 2 - i
+        if (idx < 0) break
+        const curr = currentData[idx]
+        const prev = prevData[idx]
+
+        if (
+          !curr ||
+          !prev ||
+          curr.time !== prev.time ||
+          Math.abs(curr.value - prev.value) > 0.0001
+        ) {
+          onlyLastChanged = false
+          break
+        }
+      }
+
+      if (onlyLastChanged) {
+        try {
+          // Safe to use update() - only the last point changed
+          series.update(lastCurr)
+          return true
+        } catch (e) {
+          // Fall back to setData if update fails
+          console.warn('update() failed, falling back to setData:', e)
+          series.setData(currentData)
+          return true
+        }
+      }
+    }
+
+    // Check if data is identical (no changes needed)
+    let identical = true
+    for (let i = 0; i < Math.min(10, currentData.length); i++) {
+      const idx = currentData.length - 1 - i
+      const curr = currentData[idx]
+      const prev = prevData[idx]
+      if (
+        !curr ||
+        !prev ||
+        curr.time !== prev.time ||
+        Math.abs(curr.value - prev.value) > 0.0001
+      ) {
+        identical = false
+        break
+      }
+    }
+
+    if (identical) {
+      return false // No changes detected
+    }
+  }
+
+  // For any other changes (data in middle changed, length decreased, etc.)
+  // we must use setData()
+  series.setData(currentData)
+  return true
 }
