--- apps/strength/charts/lib/primitives/CLAUDE.md
@@ -1,1 +1,0 @@
-@AGENTS.md
