+++ apps/strength/charts/lib/primitives/timeMarkers.ts
@@ -7,8 +7,6 @@
  */
 
 import { VerticalLineOptions } from './VerticalLinePrimitive'
-import { TimeRangeConfig } from './TimeRangeHighlight'
-import { COLORS } from '../../constants'
 
 export interface TimeMarkerConfig {
   /** Unique identifier for the marker */
@@ -32,163 +30,6 @@ export interface TimeMarkerConfig {
   /** Whether to show the label */
   showLabel: boolean
 }
-
-const COLOR_I = COLORS.neutral_i
-const COLOR_ASIA = COLORS.neutral
-const COLOR_US = COLORS.neutral
-
-/**
- * Configured time markers
- *
- * Add new markers here. Times are in UTC.
- * Examples:
- * - 23:00 UTC = 17:00 US Central (CST) / 18:00 US Central (CDT)
- * - 14:30 UTC = 08:30 US Central (CST) / 09:30 US Central (CDT)
- */
-export const TIME_MARKERS: TimeMarkerConfig[] = [
-  // {
-  //   label: '6am',
-  //   utcHour: 12,
-  //   utcMinute: 0,
-  //   color: COLOR_US,
-  //   labelBackgroundColor: COLOR_US,
-  //   labelTextColor: 'white',
-  //   lineStyle: 'solid',
-  //   width: 1,
-  //   showLabel: false,
-  // },
-  {
-    label: '8am',
-    utcHour: 14,
-    utcMinute: 0,
-    color: COLOR_US,
-    labelBackgroundColor: COLOR_US,
-    labelTextColor: 'white',
-    lineStyle: 'solid',
-    width: 1,
-    showLabel: false,
-  },
-  // {
-  //   label: '9am',
-  //   utcHour: 15,
-  //   utcMinute: 0,
-  //   color: COLOR_US,
-  //   labelBackgroundColor: COLOR_US,
-  //   labelTextColor: 'white',
-  //   lineStyle: 'solid',
-  //   width: 1,
-  //   showLabel: false,
-  // },
-  // {
-  //   label: '10am',
-  //   utcHour: 16,
-  //   utcMinute: 0,
-  //   color: COLOR_US,
-  //   labelBackgroundColor: COLOR_US,
-  //   labelTextColor: 'white',
-  //   lineStyle: 'solid',
-  //   width: 1,
-  //   showLabel: false,
-  // },
-  {
-    label: '11:30am',
-    utcHour: 17,
-    utcMinute: 30,
-    color: COLOR_US,
-    labelBackgroundColor: COLOR_US,
-    labelTextColor: 'white',
-    lineStyle: 'solid',
-    width: 1,
-    showLabel: false,
-  },
-  // {
-  //   label: '12am',
-  //   utcHour: 18,
-  //   utcMinute: 0,
-  //   color: COLOR_US,
-  //   labelBackgroundColor: COLOR_US,
-  //   labelTextColor: 'white',
-  //   lineStyle: 'solid',
-  //   width: 1,
-  //   showLabel: false,
-  // },
-  // {
-  //   label: '13am',
-  //   utcHour: 19,
-  //   utcMinute: 0,
-  //   color: COLOR_US,
-  //   labelBackgroundColor: COLOR_US,
-  //   labelTextColor: 'white',
-  //   lineStyle: 'solid',
-  //   width: 1,
-  //   showLabel: false,
-  // },
-  // AFTER US MARKET
-  {
-    label: '6pm',
-    utcHour: 0,
-    utcMinute: 0,
-    color: COLOR_ASIA,
-    labelBackgroundColor: COLOR_ASIA,
-    labelTextColor: 'white',
-    lineStyle: 'solid',
-    width: 1,
-    showLabel: false,
-  },
-  {
-    label: '7pm',
-    utcHour: 1,
-    utcMinute: 0,
-    color: COLOR_ASIA,
-    labelBackgroundColor: COLOR_ASIA,
-    labelTextColor: 'white',
-    lineStyle: 'solid',
-    width: 1,
-    showLabel: false,
-  },
-  {
-    label: '8pm',
-    utcHour: 2,
-    utcMinute: 0,
-    color: COLOR_ASIA,
-    labelBackgroundColor: COLOR_ASIA,
-    labelTextColor: 'white',
-    lineStyle: 'solid',
-    width: 1,
-    showLabel: false,
-  },
-]
-
-/**
- * Time Range Highlights Configuration
- *
- * Defines shaded time ranges to display on the chart.
- * For example, to highlight market hours vs. overnight sessions.
- * Times are in UTC.
- */
-export const TIME_RANGE_HIGHLIGHTS: TimeRangeConfig[] = [
-  {
-    startUtcHour: 20, // 2:45pm
-    startUtcMinute: 45,
-    endUtcHour: 13, // 7:30am
-    endUtcMinute: 30,
-    color: COLOR_I,
-  },
-  {
-    startUtcHour: 22, // 5pm
-    startUtcMinute: 0,
-    endUtcHour: 12, // 6am
-    endUtcMinute: 0,
-    color: COLOR_I,
-  },
-  {
-    startUtcHour: 3, // 9:30pm
-    startUtcMinute: 30,
-    endUtcHour: 10, // 4:30am
-    endUtcMinute: 30,
-    color: COLOR_I,
-  },
-]
 
 /**
  * Convert a UTC time to a Unix timestamp for today (or the most recent occurrence)
@@ -255,4 +96,3 @@ export function markerConfigToOptions(
     lineStyle: config.lineStyle,
   }
 }
-
